% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaggedIndividuals.R
\name{flaggedIndividuals}
\alias{flaggedIndividuals}
\title{Gives flagged individuals in a tablular format}
\usage{
flaggedIndividuals(
  model,
  data,
  individual_id,
  covariates = NULL,
  threshold = 2
)
}
\arguments{
\item{model}{An object of class "MetaboVariation"}

\item{data}{An object of class data.frame (or one that can be coerced to that class) containing data of all variables used for modelling.}

\item{individual_id}{A character string that denotes the column name of individual ids present in the dataset passed.}

\item{covariates}{A list of column names present in the data that denotes the covariates data for the individuals in the data.}

\item{threshold}{A number to filter out individuals that have been flagged in less than "threshold" metabolites. By default, it is set at 2.}
}
\value{
A data.frame of the flagged individuals that have values outside of 95\% credible interval.
}
\description{
From the class "MetaboVariation", the results contain the individuals that have been flagged as they have values outside the 95\% central credible interval based on the results.
This function provides the details of those individuals in a tabular format.
}
\examples{
\dontrun{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)
covariates = c("SexM.1F.2","Age","BMI")
individual_id = "Sample.Id"
model = MetaboVariation(data = metabol.data,individual_ids = individual_id
,metabolite = metabolites[1:3],covariates = covariates,full_posterior = TRUE)
flaggedIndividuals(model,data = metabol.data,
individual_id = individual_id,covariates=covariates)
}
}
