% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmetabolites.R
\name{getmetabolites}
\alias{getmetabolites}
\title{Extract names of metabolites present in data}
\usage{
getmetabolites(list, divider = "_", start = TRUE)
}
\arguments{
\item{list}{A list containing the name of columns that contain metabolite values across time points.}

\item{divider}{A character separating the metabolite name and time point name. The default is "_"}

\item{start}{A binary value indicating if the metabolite name is followed by the time point or if the time point is followed by metabolite.
For example, column name such as "MetaboliteA_1" should have start=TRUE, while if column name like "1_MetaboliteA" should have start=FALSE.}
}
\value{
A list containing the names of unique metabolites
}
\description{
Identify the unique metabolites present in the data. When metabolites are present at different time points, the function returns a list of unique metabolites.
}
\examples{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = getmetabolites(list = metabolite_list)
metabolites
}
