% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite.plot.R
\name{metabolite.plot}
\alias{metabolite.plot}
\title{Plots the distribution of metabolite levels.}
\usage{
metabolite.plot(data, metabolite, main = "")
}
\arguments{
\item{data}{A data frame containing data on all individuals}

\item{metabolite}{A string or a list of strings containing the names of the metabolites to be plotted.}

\item{main}{The text for the main title.}
}
\value{
A violin plot for the specified metabolites.
}
\description{
A function that plots (as a violin plot) the distributions of metabolites for all individuals across all time points.
}
\examples{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)
metabolite.plot(data = metabol.data,metabolite = metabolites[1])
metabolite.plot(data = metabol.data,metabolite = metabolites[1:3])

}
