% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolitepair.heatmap.R
\name{metabolitepair.heatmap}
\alias{metabolitepair.heatmap}
\title{Pair-wise metabolite heatmap.}
\usage{
metabolitepair.heatmap(model, interval, threshold)
}
\arguments{
\item{model}{An object of class \code{\link{MetaboVariation}} containing the fitted model results.}

\item{interval}{The interval for the highest posterior distribution (HPD) that you want to visualize.}

\item{threshold}{A positive integer indicating the minimum number of metabolites an individual must be flagged in a time point to be included in the heatmap.}
}
\value{
Returns a heatmap that illustrates the number of individuals flagged in the pair-wise metabolites within a time point.
}
\description{
Plots a heatmap that shows the number of flagged individuals in which the pair of metabolites contributes to flagging individuals within a time point.
}
\examples{
\dontrun{
# Load the simulated data and extract the metabolites names.
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)
covariates = c("SexM.1F.2","Age","BMI")
individual_id = "Individual_id"

# Run the MetaboVariation.
model = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1:3], covariates = covariates,cutoff=c(0.95,0.975,0.99))

# Visualise the metabolite pair heatmap.
metabolitepair.heatmap(model,interval=0.975,cutoff = 3)
}
}
