% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metpair.heatmap.R
\name{metpair.heatmap}
\alias{metpair.heatmap}
\title{Pair-wise metabolite heatmap.}
\usage{
metpair.heatmap(model, interval.width = 0.95)
}
\arguments{
\item{model}{An object of class \code{\link{MetaboVariation}} containing the fitted model results.}

\item{interval.width}{The width of the highest posterior density (HPD) interval considered. Must be a numeric between 0 and 1 with default value of 0.95.}
}
\value{
Returns a heatmap that illustrates the number of individuals flagged in the pairs of metabolites across all time points.
}
\description{
Plots a heatmap that shows the number of flagged individuals in which pairs of metabolites contribute to flagging the individuals across all time points.
}
\examples{
\dontrun{
# Load the simulated data and extract the metabolites names.
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)
covariates = c("SexM.1F.2","Age","BMI")
individual_id = "Individual_id"

# Run MetaboVariation on first three metabolites.
model = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1:3], covariates = covariates,interval.width=c(0.95,0.975,0.99))

# Visualise the metabolite pair heatmap.
metpair.heatmap(model,interval.width=0.975)
}
}
