% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetabolite.R
\name{plotMetabolite}
\alias{plotMetabolite}
\title{Initiate a data distribution visualization for the given metabolite.}
\usage{
plotMetabolite(data, metabolite, points = FALSE)
}
\arguments{
\item{data}{An object of class data.frame (or one that can be coerced to that class) containing data of all variables used in the plot}

\item{metabolite}{A character or a list of characters containing metabolites name for which the plot will be drawn. If a single metabolite is passed
then a plot for that metabolite is drawn. If a list of metabolite is passed then each metabolite will be shown in a different plot.}

\item{points}{As violin plot shows the distribution, observations along the end of tail are also shown by default. If you want to remove these points, then set points = FALSE.}
}
\value{
It returns plot for the metabolites passed.
}
\description{
It shows a violin plot for different timepoints for the passed metabolites. Violin plot is a plot that shows a distribution of the
numeric list.
}
\examples{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)
plotMetabolite(data = metabol.data,metabolite = metabolites[1])
plotMetabolite(data = metabol.data,metabolite = metabolites[1:3],points = TRUE)

}
