% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.MetaboVariation.R
\name{Shiny.MetaboVariation}
\alias{Shiny.MetaboVariation}
\title{MetaboVariation Shiny app launcher}
\usage{
Shiny.MetaboVariation()
}
\description{
Launches the Shiny application for the MetaboVariation functionality. The
interactive application allows to analyse data and flag individuals that have variations in their metabolite levels.
}
