% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.plot.R
\name{circos.plot}
\alias{circos.plot}
\title{Plots the metabolite-specific circos plot.}
\usage{
circos.plot(model, metabolite, interval.width = 0.95, title = "")
}
\arguments{
\item{model}{An object of class \code{\link{MetaboVariation}} containing the fitted model results.}

\item{metabolite}{A string containing the name of the metabolite you want to visualise.}

\item{interval.width}{The width of the highest posterior density (HPD) interval considered. Must be a numeric between 0 and 1 with default value of 0.95.}

\item{title}{Text for the title of the plot.}
}
\description{
Visualise a circos plot showing a highest posterior density (HPD) interval for posterior predictive distributions for all individuals across all time points for a specific metabolite.
}
\examples{
\dontrun{
# Load the simulated data and extract the metabolites names.
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)
covariates = c("SexM.1F.2","Age","BMI")
individual_id = "Individual_id"

# Run MetaboVariation on first three metabolites.
model = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1:3], covariates = covariates,interval.width=c(0.95,0.975,0.99))

# Plots a circos plot for first metabolite with an interval width of 0.975.
circos.plot(model,metabolite = metabolites[1],interval.width=0.975)
}
}
