% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaggedIndividuals.R
\name{flaggedIndividuals}
\alias{flaggedIndividuals}
\title{Lists flagged individuals in tabular format.}
\usage{
flaggedIndividuals(model, data, individual_id, threshold = 2)
}
\arguments{
\item{model}{An object of class \code{\link{MetaboVariation}}.}

\item{data}{A data frame containing data of all variables used in the BGLM along with a column detailing the individual ids.}

\item{individual_id}{A character string detailing the name of the column in the data frame that contains the individual ids.}

\item{threshold}{A positive integer. Individuals that have been flagged in less than \emph{threshold} times across different metabolites will not be included. By default, it is set to 2.}
}
\value{
The function will provide two types of lists based on the type of model object.
\itemize{
\item Single metabolite model - If the model contains result for single metabolite, the function will return a list of flagged individuals along with their metabolite levels across different time points.
\item Multiple metabolites model - If the model contain results for multiple metabolites, the function will return a list of individuals flagged in those metabolites  along with the number of times that individual have been flagged in each metabolite.
}
}
\description{
A function to list the flagged individuals resulting from fitting a Bayesian generalised linear model (BGLM) model to repeated measures data.
}
\examples{
\dontrun{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = getmetabolites(list = metabolite_list)
individual_id = "Individual_id"
model = MetaboVariation(data = metabol.data,individual_ids = individual_id
,metabolite = metabolites,full_posterior = TRUE)
flaggedIndividuals(model,data = metabol.data,
individual_id = individual_id)
}
}
\seealso{
\code{\link{MetaboVariation}}
}
