% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.metabolites.R
\name{get.metabolites}
\alias{get.metabolites}
\title{Extract metabolite names present in data}
\usage{
get.metabolites(list, divider = "_", start = TRUE)
}
\arguments{
\item{list}{A list containing names of columns that store metabolites values across timepoints.}

\item{divider}{A character value that separates the metabolite name and timepoint value. The default is "_"}

\item{start}{A binary value that tells the function if the metabolite name is followed by timepoint or the timepoint is followed by metabolite.
For eg: With divider "_", column names in format "MetaboliteA_1" should have start=TRUE, while column names in format "1_MetaboliteA" should have start=FALSE.}
}
\value{
A list of unique metabolite names
}
\description{
get_metabolites() is a function that will extract the unique metabolites names from the columns present in data, when different timepoints of metabolites are listed as different columns in data.
}
\examples{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)
metabolites
}
