% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite.plot.R
\name{metabolite.plot}
\alias{metabolite.plot}
\title{Plot the distribution of metabolite levels.}
\usage{
metabolite.plot(data, metabolite, title = "")
}
\arguments{
\item{data}{A data frame containing data of all variables to be used in the BGLM. Refer to \code{\link{metabol.data}} for the required structure and content of the data frame.}

\item{metabolite}{A string or a list of strings containing the names of the metabolites to be plotted.}

\item{title}{The text for the title of the plot.}
}
\description{
A function that produces violin plots of the distributions of metabolites for all individuals across all time points.
}
\examples{
# Load the simulated data and extract the metabolites names.
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)

# Plots the distribution of a single metabolite or multiple metabolites.
metabolite.plot(data = metabol.data,metabolite = metabolites[1])
metabolite.plot(data = metabol.data,metabolite = metabolites[1:3])

}
