% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboVariation.R
\name{MetaboVariation}
\alias{MetaboVariation}
\title{Fits a Bayesian generalised linear model to repeated measurements of metabolites.}
\usage{
MetaboVariation(
  data,
  individual_ids,
  metabolite,
  covariates = NULL,
  save_brms_model = FALSE,
  full_posterior = FALSE,
  iter = 2000,
  warmup = floor(iter/2),
  thin = 1,
  cutoff = 95,
  ...
)
}
\arguments{
\item{data}{A data frame containing data of all variables to be used in the BGLM.}

\item{individual_ids}{A character string detailing the name of the column in the data that contains the individual ids.}

\item{metabolite}{A string or vector of string containing the name of metabolites to model.}

\item{covariates}{A list of string containing the name of columns in the data that contains the covariates.}

\item{save_brms_model}{A logical value indicating whether the model should be saved or not. By default, it is set as FALSE.}

\item{full_posterior}{A logical value indicating whether the full posterior distribution should be returned or not. By default, it is set as FALSE. #'}

\item{iter}{Number of total iterations per chain (including warmup; defaults to 2000).}

\item{warmup}{A positive integer specifying number of warmup (aka burnin)
iterations. This also specifies the number of iterations used for stepsize
adaptation, so warmup draws should not be used for inference. The number
of warmup should not be larger than \code{iter} and the default is
\code{iter/2}.}

\item{thin}{Thinning rate. Must be a positive integer. Set thin > 1 to save memory and computation time if iter is large.}

\item{cutoff}{Cutoff for the credible interval of posterior predictive distribution. The value should be less than 100. Default is set to 95.}

\item{...}{Further arguments are passed to \code{\link[brms]{brm}} for backend processes.}
}
\value{
function returns an object of "MetaboVariation". Following values are returned for each metabolite object.
\itemize{
\item significant_covariates - a list of covariates that influences the metabolite value significantly for the individuals.
\item result - A summarized result for the posterior predictive distribution of the cohort. It contains the mean, \code{cutoff}\% credible interval range, tails of the credible interval and the original value of the individual for that timepoint. It also has a flag value that shows whether the original value lies within the credible interval or not.
\item iter - Number of total iterations per chain used for modelling.
\item warmup - Number of total warmup iterations per chain.
\item Rhat - Rhat refers to the potential scale reduction statistic, also known as the Gelman-Rubin statistic.
Thus, it measures the extent to which chains are converging. The further the value of the statistic from 1, the worse. Model is considered good if the value lies between 0.9 and 1.05.
\item metabolite - Contains the name of metabolite on which modelling is done.
\item full_posterior - It contains the full posterior predictive distribution of the data. It can be used for further analysis as per the user requirements.
}
}
\description{
Implements a Bayesian generalised linear model (BGLM) to identify the flagged individuals with notable variations.
}
\details{
TheBayesian generalised linear model (BGLM) to identify individuals with variations in their
metabolite levels is fitted by employing STAN. The BGLM includes covariates  of the individuals and models the repeated measurements to explore intra-individual variation. Individuals with notable variation in metabolite levels are flagged.
The model builds on the Bayesian framework using the function \code{\link[brms:brm]{brm}} from \code{\link[brms:brm]{brm}} package.

The BGLM models the metabolite value \eqn{M_{ij}} for the individual \eqn{i} at timepoint \eqn{j},
\deqn{ M_{ij} = \beta_{0} + \beta_{1}X_i + r_i + \epsilon_{ij}}

where \eqn{i \in 1, 2, ..., n} denotes individual \eqn{i} of the \eqn{n} individuals in the cohort and \eqn{j \in 1, 2, ..., k} denotes the timepoints at which metabolites are measured. The intercept of the model is denoted by \eqn{\beta_0}, model coefficients for \eqn{m} covariates are denoted by \eqn{\beta_{1}} which is in form of \eqn{1 * m} matrix and \eqn{m} covariates of the \eqn{i^{th}} individual is represented by \eqn{X_i} in theform of the diagonal matrix of \eqn{m * m}.
The random effect due to repeated measurements of the \eqn{i^{th}} individual is represented by \eqn{r_i} and \eqn{\epsilon_{ij}} indicates the model's error term for individual \eqn{i} at time point \eqn{j}.
}
\examples{
\dontrun{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = getmetabolites(list = metabolite_list)
covariates = c("SexM.1F.2","Age","BMI")
individual_id = "Sample.Id"
model = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1], covariates = covariates)
model = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1:3], full_posterior = TRUE)
}

}
\seealso{
\code{\link{flaggedIndividuals}}, \code{\link{plot.MetaboVariation}}
}
