% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MetaboVariation}
\alias{plot.MetaboVariation}
\alias{plot}
\title{Plots the flagged individuals for the metabolites.}
\usage{
\method{plot}{MetaboVariation}(x, type = "circos", timepoints = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link{MetaboVariation}}}

\item{type}{A character specifying the plot type (eg: "circos", "metabolites_count"). By default, "circos" is shown.}

\item{timepoints}{A numeric vector containing timepoints to be shown in the plots. By default, it is NULL and will show all the timepoints}

\item{...}{Further arguments are ignored.}
}
\value{
plot will show the results of modelling in the following two ways:
\itemize{
\item Circos plot - A circular plot that will show the credible interval for all the individuals across timepoints. If the observed value
is outside the credible interval, that bar will be flagged as black and the individual's id will be shown in \strong{bold}.
\item Metabolites count plot - A bar plot that shows the number of individuals that are flagged in each timepoint for all metabolites passed.
}
}
\description{
A function to plot the results after fitting a Bayesian generalised linear model (BGLM) model to metabolomic data.
A circos plot or a barplot can be produced depicting results from different perspectives.
}
\examples{
\dontrun{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)
covariates = c("SexM.1F.2","Age","BMI")
individual_id = "Sample.Id"
model = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1:3], covariates = covariates,full_posterior = TRUE)
plot(x = model, type = "circos")
plot(x = model, type = "metabolites_count",timepoints = c(1,2,4))
plot(x = model, type = "individual_count")
}
}
\seealso{
\code{\link{MetaboVariation}}
}
