% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MetaboVariation}
\alias{plot.MetaboVariation}
\alias{plot}
\title{Plots the flagged individuals for the metabolites.}
\usage{
\method{plot}{MetaboVariation}(x, type = "circos", timepoints = NULL, main = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link{MetaboVariation}}}

\item{type}{A character specifying the plot type either \code{circos} or  \code{metabolites_count}. By default, \code{circos} is shown.}

\item{timepoints}{A numeric vector denoting the timepoints to be shown in the plots. By default, it is NULL and will show all the timepoints.}

\item{main}{Title for the plot.}

\item{...}{Further arguments are ignored.}
}
\value{
This function will show the results of modelling in the following two ways:
\itemize{
\item Circos plot - Plot showing the posterior predictive Highest Posterior Distribution (HPD) interval for all the individuals across timepoints. If the observed value
is outside the HPD interval, that bar will be flagged as black and the individual's id will be shown in \strong{bold} font.
\item Metabolites count plot - Bar plot that shows the number of individuals that are flagged in each timepoint for all metabolites in the \code{\link{MetaboVariation}} object.
}
}
\description{
A function to plot the results after fitting a Bayesian generalised linear model (BGLM) model to data.
Two types of plots can be produced depicting results from different perspectives.
}
\examples{
\dontrun{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)
covariates = c("SexM.1F.2","Age","BMI")
individual_id = "Individual_id"
model = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1:3], covariates = covariates,full_posterior = TRUE)
plot(x = model, type = "circos")
plot(x = model, type = "metabolites_count",timepoints = c(1,2,4))
}
}
\seealso{
\code{\link{MetaboVariation}}
}
