% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetabolite.R
\name{plotMetabolite}
\alias{plotMetabolite}
\title{Plots the distribution of metabolite.}
\usage{
plotMetabolite(data, metabolite)
}
\arguments{
\item{data}{A data frame containing data of all variables.}

\item{metabolite}{A string or a list of string containing names of metabolites to plot.}
}
\value{
It returns a violin plot for the metabolites passed.
}
\description{
A function that displays the distribution of a metabolite levels for all individuals across all time points as a violin plot.
}
\examples{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = getmetabolites(list = metabolite_list)
plotMetabolite(data = metabol.data,metabolite = metabolites[1])
plotMetabolite(data = metabol.data,metabolite = metabolites[1:3])

}
