% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetabolite.R
\name{plotMetabolite}
\alias{plotMetabolite}
\title{Plots the distribution of metabolite.}
\usage{
plotMetabolite(data, metabolite, main = "")
}
\arguments{
\item{data}{A data frame containing data of all variables.}

\item{metabolite}{A string or a list of string containing names of metabolites to plot.}

\item{main}{the text for the main title.}
}
\value{
It returns a violin plot for the given metabolites.
}
\description{
A function that plots the distributions of metabolites for all individuals across all time points. It can plot for either a single metabolite or multiple metabolites.
}
\examples{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = getmetabolites(list = metabolite_list)
plotMetabolite(data = metabol.data,metabolite = metabolites[1])
plotMetabolite(data = metabol.data,metabolite = metabolites[1:3])

}
