% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Predict method for mixed models fitted with MCMCglmm in MetaboVariation.}
\usage{
prediction(
  object,
  seed = NULL,
  nsim = nrow(object$Sol),
  newdata = NULL,
  level = 0.95,
  verbose = FALSE
)
}
\arguments{
\item{object}{An object of class \code{\link{MetaboVariation}}.}

\item{seed}{The seed for random number generation to make results reproducible. A seed is default to 19205033.}

\item{nsim}{number of response vectors to simulate. Defaults to number of effective sample size of the model.}

\item{newdata}{An optional data.frame for which to evaluate predictions. If NULL (default), the original data of the model is used.}

\item{level}{A numeric scalar in the interval (0,1) giving the target probability content of the intervals.}

\item{verbose}{logical; if TRUE, warnings are issued with newdata when the original model has fixed effects that do not appear in newdata and/or newdata has random effects not present in the original model.}
}
\value{
A matrix of expectated values and Highest Posterior Distribution interval.
}
\description{
Compute posterior draws of the posterior predictive distribution. Can be performed for the data used to fit the model (posterior predictive checks) or for new data.
It is similar to the function \code{\link[MCMCglmm:predict.MCMCglmm]{predict}} from \code{\link[MCMCglmm]{MCMCglmm}} package.
}
\details{
Title
}
