% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_check.R
\name{predictive_check}
\alias{predictive_check}
\title{Performs posterior predictive check for two MetaboVariation objects}
\usage{
predictive_check(
  data,
  model1,
  model2,
  timepoints,
  metabolites,
  replication = 100
)
}
\arguments{
\item{data}{A data frame containing data of all variables to be used in the BGLM.}

\item{model1}{An object of class \code{\link{MetaboVariation}} containing the fitted model results.}

\item{model2}{A different object of class \code{\link{MetaboVariation}} containing the fitted model results.}

\item{timepoints}{A list of time points present in the data.}

\item{metabolites}{A string or vector of string containing the name of metabolites to model.}

\item{replication}{number of replicated data set to make. Defaults to 100. Should be less than effective sample size.}
}
\value{
A boxplot comparing the differences between two \code{\link{MetaboVariation}} objects. Replicated datasets are created from both models,
generating a correlation matrix of the original dataset and replicated dataset for each timepoint. The boxplot displays the mean absolute difference for each timepoint.
}
\description{
Performs posterior predictive check for two MetaboVariation objects
}
\examples{
\dontrun{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)
covariates = c("SexM.1F.2","Age","BMI")
individual_id = "Individual_id"
model = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1], covariates = covariates)
model1 = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1:3],type="dependent")
model2 = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1:3],type="independent")
predictive_check(metabol.data, model1, model2, c(1,2,3,4), metabolites)
}
}
