% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar.plot.R
\name{radar.plot}
\alias{radar.plot}
\title{Plot a radarplot showing metabolic profile of an individual across all timepoints.}
\usage{
radar.plot(model, interval, individual, title = "")
}
\arguments{
\item{model}{An object of class \code{\link{MetaboVariation}} containing the fitted model results.}

\item{interval}{The interval for the Highest Posterior Distribution (HPD) that you want to visualize.}

\item{individual}{ID for individual you want to visualize}

\item{title}{Title to plot}
}
\value{
Returns a radarplot for an specific individual at specified interval
}
\description{
Plot a radarplot showing metabolic profile of an individual across all timepoints.
}
\examples{
\dontrun{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)
covariates = c("SexM.1F.2","Age","BMI")
individual_id = "Individual_id"
model = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1:3], covariates = covariates,cutoff=c(0.95,0.975,0.99))
radarplor(model,interval=0.975,individual = 3)
}
}
