% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.MetaboVariation.R
\name{Shiny.MetaboVariation}
\alias{Shiny.MetaboVariation}
\title{Fits a Bayesian generalised linear model to repeated measurements of metabolites.}
\usage{
Shiny.MetaboVariation(
  data,
  individual_ids,
  metabolite,
  covariates = NULL,
  save_brms_model = FALSE,
  full_posterior = FALSE,
  iter = 2000,
  warmup = floor(iter/2),
  thin = 1,
  cutoff = 95,
  ...
)
}
\arguments{
\item{data}{A data frame containing data of all variables to be used in the BGLM.}

\item{individual_ids}{A character string detailing the name of the column in the data that contains the individual ids.}

\item{metabolite}{A string or vector of string containing the name of metabolites to model.}

\item{covariates}{A list of string containing the name of columns in the data that contains the covariates.}

\item{save_brms_model}{A logical value indicating whether the model should be saved or not. By default, it is set as FALSE.}

\item{full_posterior}{A logical value indicating whether the full posterior distribution should be returned or not. By default, it is set as FALSE.}

\item{iter}{Number of total iterations per chain (including warmup; defaults to 2000).}

\item{warmup}{A positive integer specifying number of warmup (aka burnin)
iterations. This also specifies the number of iterations used for stepsize
adaptation, so warmup draws should not be used for inference. The number
of warmup should not be larger than \code{iter} and the default is
\code{iter/2}.}

\item{thin}{Thinning rate. Must be a positive integer. Set thin > 1 to save memory and computation time if iter is large.}

\item{cutoff}{Cutoff for the credible interval of posterior predictive distribution. The value should be less than 100. Default is set to 95.}

\item{...}{Further arguments are passed to \code{\link[brms]{brm}} for backend processes.}
}
\value{
function returns an object of "MetaboVariation". Following values are returned for each metabolite object.
\itemize{
\item metabolite - contains the name of metabolite.
\item significant_covariates - a list of covariates that have significant relationship (at the 90\% level) with the metabolite levels in individuals.
\item result - a summary of the posterior predictive distribution of the cohort. The result contains the mean, \code{cutoff} \% credible interval width, tails of the credible interval and the observed value of the metabolite for the individual for that timepoint. The result also has a binary flag that shows whether the observed value lies within the credible interval or not.
\item warmup - number of total warmup iterations per chain.
\item iter - number of total iterations per chain used for modelling.
\item Rhat - the potential scale reduction statistic, also known as the Gelman-Rubin statistic which measures the extent to which chains are converging. The further the value of the statistic from 1, the poorer the convergence of the chains. The MCMC chains are considered converged if the value lies between 0.9 and 1.05.
\item full_posterior - it contains the full posterior predictive distribution of the data.
}
}
\description{
Implements a Bayesian generalised linear model (BGLM) to identify the flagged individuals with notable variations.
}
\details{
The Bayesian generalised linear model (BGLM) to identify individuals with variations in their
metabolite levels is fitted by employing STAN. The BGLM can include covariates and learns the intra-individual variation in metabolite levels using repeated measures.
Posterior predictive distributions of metabolite levels at the individual level are provided. These are used to identify individuals with observed metabolite level outside the
\code{cutoff}\% central credible interval at one time point; such individuals with notable variation are flagged.
The model builds on the Bayesian framework using the function \code{\link[brms:brm]{brm}} from \code{\link[brms:brm]{brms}} package.

The BGLM models the metabolite value \eqn{M_{ij}} for the individual \eqn{i} at timepoint \eqn{j},
\deqn{ y^m_{it} = \beta_{0} + \sum_{l=1}^L\beta_{1,l}x_{il} + S_{it} + \epsilon_{it}}

where \eqn{y^m_{it}} is the metabolite level for individual \eqn{i} at time point \eqn{t}, the term \eqn{X_{il}} is the covariate \eqn{l \in 1, 2, ..., L} for individual \eqn{i}. The notation \eqn{\beta_0} is the mean intercept, while \eqn{\beta_{1,l}} is the regression coefficient for the covariate \eqn{l}. The random effect for the \eqn{i^{th}} individual is denoted by \eqn{S_{it}} for timepoint \eqn{t} and \eqn{\epsilon_{it}} is the random error for individual \eqn{i} at timpoint \eqn{t} assumed \eqn{\epsilon_{ij} \sim N(0,\sigma^2)}.
}
\examples{
\dontrun{
data(metabol.data)
metabolite_list = colnames(metabol.data)[5:length(colnames(metabol.data))]
metabolites = get.metabolites(list = metabolite_list)
covariates = c("SexM.1F.2","Age","BMI")
individual_id = "Individual_id"
model = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1], covariates = covariates)
model = MetaboVariation(data = metabol.data,individual_ids = individual_id,
metabolite = metabolites[1:3], full_posterior = TRUE)
}

}
\seealso{
\code{\link{flagged.Individuals}}, \code{\link{plot.MetaboVariation}}
}
